smartcap-ai/requirements.txt
fastapi
uvicorn
opencv-python
numpy
python-multipart
starlette
websockets
yarg==0.1.9
torch
numpy
scipy
lap
ultralytics
cython_bbox
redis
python-dotenv
httpx
